Modified beans
- Change byte for int in the identifiers
- Set static to the classes used for complex index
- Set fetch lazy

+++ src/main/java/es/a2m/calendar/domain/Ausencia.java
@@ -22,7 +22,7 @@ public class Ausencia implements Serializable {
 	@Id
 	@GeneratedValue
 	@Column(name="COD_AUSENCIA", nullable=false)
-	private byte codAusencia;
+	private int codAusencia;
 
 	@Column(name="DESCRIPCION", nullable=false, length=50)
 	private String descripcion;
